+++ src/main/java/com/inventory/engine/repository/ProductRepository.java
@@ -6,6 +6,6 @@ import java.util.Optional;
 
 public interface ProductRepository {
     void save(Product product);
-    Optional<Product>findById(int id);
+    Optional<Product>findById(Long id);
     List<Product>findAll();
 }
